api-tests: reconfigure to bot inherit from project parent
The parent would pull in surefire and failsafe configuration that wouldn’t allow native testing to find the native image. Thus we needed to add in those elements from parent that we still needed, such as the compiler configuration.

The native profile is used for testing against the generated native image. Without it, testing is carried out as normal against the jar files.