Fix native image (#187)
* reflection-config.json MUST be in runner module

When it is in a dependency it is stored within that dependency jar within the internal lib directory and the runner doesn’t see it on the bootstrap classpath.

* runner: remove unneeded reflection-config.json

* entities.Veterinarians: add setters

* entities.Owner: add setters

* entities.Pet: add setters

* entities.Visit: add setters

* Makefile: add native and run-native targets

* github actions: add native-image job

* github actions: build native image

* github actions: native build needs checkout

* api-tests.NativeVisitsIT: added

* api-tests.Native{Owners,Pets,Vets}IT: added

* api-tests: rename non-native test classes as *ApiTest

* api-tests: reconfigure to bot inherit from project parent

The parent would pull in surefire and failsafe configuration that wouldn’t allow native testing to find the native image. Thus we needed to add in those elements from parent that we still needed, such as the compiler configuration.

The native profile is used for testing against the generated native image. Without it, testing is carried out as normal against the jar files.

* api-tests: run the native-image tests

* github actions: configure postgres service

* github actions: specify db name to create